@@ -1,19 +1,14 @@
 import React from 'react'
-import {
-  BrowserRouter as Router,
-  Switch,
-  Route,
-  Redirect,
-} from 'react-router-dom'
+import { Route, BrowserRouter, Routes } from 'react-router-dom'
 
 import { Navbar } from './app/Navbar'
 
 function App() {
   return (
-    <Router>
+    <BrowserRouter>
       <Navbar />
       <div className="App">
-        <Switch>
+        <Routes>
           <Route
             exact
             path="/"
@@ -23,10 +18,9 @@ function App() {
               </section>
             )}
           />
-          <Redirect to="/" />
-        </Switch>
+        </Routes>
       </div>
-    </Router>
+    </BrowserRouter>
   )
 }
 
